root_dir_classes_app/requirements.txt
about-time==4.2.1
alive-progress==3.3.0
asttokens==3.0.1
attrs==25.4.0
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
graphemeu==0.7.2
import-ipynb==0.2
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
rpds-py==0.30.0
scikit-learn==1.8.0
scikit-surprise==1.1.4
scipy==1.16.3
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.2.14
